% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSO_design_TE.gbop2.R
\name{PSO_design_TE}
\alias{PSO_design_TE}
\title{PSO: Optimal/Minimax design with toxicity and efficacy boundaries}
\usage{
PSO_design_TE(
  design = "optimal",
  unified.u = 1,
  method = "quantum",
  nlooks = 4,
  skip_efficacy = NULL,
  skip_toxicity = NULL,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  e1n = 0.3,
  e2n = 0.4,
  e3n = 0.2,
  e1a = 0.6,
  e2a = 0.2,
  e3a = 0.15,
  err_eff = 0.1,
  err_tox = 0.1,
  err_all = 0.05,
  power_eff = 0.8,
  power_tox = 0.8,
  power_all = 0.8,
  seed = 1024,
  nSwarm = 32,
  maxIter = 100
)
}
\arguments{
\item{design}{choose from "optimal", "minimax", or "unified"}

\item{unified.u}{specify when design = "unified", u in zero to one}

\item{method}{method for single PSO, choose from "default", "quantum" or "dexp"}

\item{nlooks}{number of interim looks}

\item{skip_efficacy}{default is NULL, indicate skip efficacy as 1 and not skip as 0 in a vector}

\item{skip_toxicity}{default is NULL, indicate skip toxicity as 1 and not skip as 0 in a vector}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{e1n}{H0 for efficacy}

\item{e2n}{H0 for toxicity}

\item{e3n}{H0 for Eff and Tox}

\item{e1a}{H1 for efficacy}

\item{e2a}{H1 for toxicity}

\item{e3a}{H1 for Eff and Tox}

\item{err_eff}{Type I error rate: Efficacious but toxic}

\item{err_tox}{Type I error rate: Safe but futile}

\item{err_all}{Type I error rate: Futile and toxic}

\item{power_eff}{power: Efficacious but toxic}

\item{power_tox}{power: Safe but futile}

\item{power_all}{power: Futile and toxic}

\item{seed}{Random seed for reproducibility}

\item{nSwarm}{nSwarm in PSO}

\item{maxIter}{maxIter in PSO}
}
\value{
A list on design parameters and operating characteristics
}
\description{
This function implements Particle Swarm Optimization (PSO) to find an optimal or minimax design with toxicity and efficacy boundaries.
}
\examples{
PSO_design_TE(design = "optimal",
unified.u = 1,
method = "quantum",
nlooks = 1,
skip_efficacy = NULL, 
skip_toxicity = NULL , ## Default:NULL, 1 is skip.
maxPatients = 20, ## maximum number of patients
Nmin_cohort1 = 10,
Nmin_increase = 5,
e1n = 0.3,  # H0 for Eff
e2n = 0.4,  # H0 for Tox
e3n = 0.2, # H0 for Eff and Tox
e1a = 0.6,  # Ha for Eff
e2a = 0.2,  # Ha for Tox
e3a = 0.15, # Ha for Eff and Tox
err_eff = 0.1,  # Type I error rate: Efficacious but toxic
err_tox = 0.1,  # Type I error rate: Safe but futile
err_all = 0.05,  # Type I error rate: Futile and toxic
power_eff = 0.8,
power_tox = 0.8,
power_all = 0.8,
seed = 1024,
nSwarm = 32,
maxIter = 100)


}
