% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSO_power.gbop2.R
\name{PSO_power}
\alias{PSO_power}
\title{PSO maximizing power design with single boundary}
\usage{
PSO_power(
  nlooks = 1,
  totalPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  method = "default",
  b1n = 0.2,
  b1a = 0.4,
  err1 = 0.05,
  minPower = 0.8,
  seed = 1024,
  nSwarm = 64,
  maxIter = 200
)
}
\arguments{
\item{nlooks}{number of looks}

\item{totalPatients}{total number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{method}{method for single PSO, choose from "default", "quantum" or "dexp"}

\item{b1n}{Null hypothesis response rate}

\item{b1a}{Alternative hypothesis response rate}

\item{err1}{Type I error rate}

\item{minPower}{power}

\item{seed}{seed for pso}

\item{nSwarm}{nSwarm for pso}

\item{maxIter}{maxIter for pso}
}
\value{
A list
}
\description{
PSO maximizing power design with single boundary
}
\examples{

 PSO_power(
 nlooks = 1,
 totalPatients = 50, ## total patients in optimal_power
 Nmin_cohort1 = 10,
 Nmin_increase = 5,
 method = "default", # "quantum", "dexp"
 b1n = 0.2,  # Null hypothesis response rate
 b1a = 0.4,  # Alternative hypothesis response rate
 err1 = 0.05,  # Type I error rate
 minPower = 0.8, ## power
 seed = 1024,   ## set seed to calculate OC
 nSwarm = 64,
 maxIter = 200
 )
}
