% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOdesign.gbop2.R
\name{PSOdesign}
\alias{PSOdesign}
\title{PSO for optimal/minimax design with single boundary}
\usage{
PSOdesign(
  design = "optimal",
  nlooks = 1,
  b1n = 0.2,
  b1a = 0.4,
  err1 = 0.05,
  minPower = 0.8,
  weight = 1,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  method = "default",
  seed = 1024,
  nSwarm = 64,
  maxIter = 200
)
}
\arguments{
\item{design}{choose from "optimal" or "minimax"}

\item{nlooks}{number of interim looks}

\item{b1n}{Null hypothesis response rate}

\item{b1a}{Alternative hypothesis response rate}

\item{err1}{Type I error rate}

\item{minPower}{power}

\item{weight}{weight of sample size under null}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{method}{method for single PSO, choose from "default", "quantum" or "dexp"}

\item{seed}{seed for pso}

\item{nSwarm}{nSwarm for pso}

\item{maxIter}{maxIter for pso}
}
\value{
A list
}
\description{
PSO for optimal/minimax design with single boundary
}
