% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSODesign_TE.gbop2.R
\name{PSODesign_TE}
\alias{PSODesign_TE}
\title{PSO for optimal/minimax design with efficacy and toxicity boundaries}
\usage{
PSODesign_TE(
  design = "optimal",
  method = "quantum",
  nlooks = 2,
  skip_efficacy = NULL,
  skip_toxicity = NULL,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  e1n = 0.3,
  e2n = 0.4,
  e3n = 0.2,
  e1a = 0.6,
  e2a = 0.2,
  e3a = 0.15,
  err_eff = 0.1,
  err_tox = 0.1,
  err_all = 0.05,
  power_eff = 0.8,
  power_tox = 0.8,
  power_all = 0.8,
  nSwarm = 32,
  maxIter = 100
)
}
\arguments{
\item{design}{choose from "optimal" or "minimax"}

\item{method}{method for single PSO, choose from "default", "quantum" or "dexp"}

\item{nlooks}{number of interim looks}

\item{skip_efficacy}{default is NULL, indicate skip efficacy as 1 and not skip as 0 in a vector}

\item{skip_toxicity}{default is NULL, indicate skip toxicity as 1 and not skip as 0 in a vector}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{e1n}{H0 for efficacy}

\item{e2n}{H0 for toxicity}

\item{e3n}{H0 for Eff and Tox}

\item{e1a}{H1 for efficacy}

\item{e2a}{H1 for toxicity}

\item{e3a}{H1 for Eff and Tox}

\item{err_eff}{Type I error rate: Efficacious but toxic}

\item{err_tox}{Type I error rate: Safe but futile}

\item{err_all}{Type I error rate: Futile and toxic}

\item{power_eff}{power: Efficacious but toxic}

\item{power_tox}{power: Safe but futile}

\item{power_all}{power: Futile and toxic}

\item{nSwarm}{nSwarm in PSO}

\item{maxIter}{maxIter in PSO}
}
\description{
PSO for optimal/minimax design with efficacy and toxicity boundaries
}
\examples{
PSODesign_TE(design = "optimal",
method = "quantum",
nlooks = 2,
skip_efficacy = NULL, # NULL no skipping c(), 1 is skip
skip_toxicity = NULL , ## NULL c()
maxPatients = 50, ## maximum number of patients
Nmin_cohort1 = 10,
Nmin_increase = 5,
e1n = 0.3,  # H0 for Eff
e2n = 0.4,  # H0 for Tox
e3n = 0.2, # H0 for Eff and Tox
e1a = 0.6,  # Ha for Eff
e2a = 0.2,  # Ha for Tox
e3a = 0.15, # Ha for Eff and Tox
err_eff = 0.1,  # Type I error rate: Efficacious but toxic
err_tox = 0.1,  # Type I error rate: Safe but futile
err_all = 0.05,  # Type I error rate: Futile and toxic
power_eff = 0.8,
power_tox = 0.8,
power_all = 0.8,
nSwarm = 32,
maxIter = 100)
}
