% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOGO_design.gbop2.R
\name{PSOGO_Design}
\alias{PSOGO_Design}
\title{PSOGO for optimal/minimax design with single boundary}
\usage{
PSOGO_Design(
  design = "optimal",
  nlooks = 1,
  b1n = 0.2,
  b1a = 0.4,
  err1 = 0.05,
  nParallel = 3,
  minPower = 0.8,
  weight = 1,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  pso_method = "all",
  nSwarm = 64,
  maxIter = 200,
  nCore = 4
)
}
\arguments{
\item{design}{choose from "optimal" or "minimax"}

\item{nlooks}{number of interim looks}

\item{b1n}{Null hypothesis response rate}

\item{b1a}{Alternative hypothesis response rate}

\item{err1}{Type I error rate}

\item{nParallel}{number of pso ensemble}

\item{minPower}{power}

\item{weight}{weight of sample size under null}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{pso_method}{"all" for using three distinct pso, otherwise indicate single pso method}

\item{nSwarm}{nSwarm for pso}

\item{maxIter}{maxIter for pso}

\item{nCore}{number of core}
}
<<<<<<< HEAD
=======
\value{
A list on design parameters and operating characteristics
}
>>>>>>> e2eb909070304bacfc956234fcbae6ed054f6c3c
\description{
PSOGO for optimal/minimax design with single boundary
}
\examples{
PSOGO_Design(design = "optimal", #"minimax"
nlooks = 1,
b1n = 0.2,  # Null hypothesis response rate
b1a = 0.4,  # Alternative hypothesis response rate
err1 = 0.05,  # Type I error rate
nParallel = 3,
minPower = 0.8, ## power
weight = 1, ## weight of sample size under null
maxPatients = 50,  ## maximum number of patients
Nmin_cohort1 = 10,
Nmin_increase = 5,
pso_method = "all", ## or three single pso "default", "quantum", "dexp"
nSwarm = 64,
maxIter = 200,
nCore = 4)
}
