% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSOGPower_dual.gbop2.R
\name{PSOGPower_dual}
\alias{PSOGPower_dual}
\title{PSOGO for maximizing power design with dual boundaries}
\usage{
PSOGPower_dual(
  design = "optimal",
  nlooks = 1,
  b1n = 0.2,
  b1a = 0.4,
  err1 = 0.05,
  nParallel = 3,
  minPower = 0.8,
  weight = 1,
  totalPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  pso_method = "all",
  seed = 1024,
  nSwarm = 64,
  maxIter = 200,
  nCore = 4
)
}
\arguments{
\item{design}{choose from "optimal" or "minimax"}

\item{nlooks}{number of interim looks}

\item{b1n}{Null hypothesis response rate}

\item{b1a}{Alternative hypothesis response rate}

\item{err1}{Type I error rate}

\item{nParallel}{number of pso ensemble}

\item{minPower}{power}

\item{weight}{weight of sample size under null}

\item{totalPatients}{total patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{pso_method}{"all" for using three distinct pso, otherwise indicate single pso method}

\item{seed}{seed for pso}

\item{nSwarm}{nSwarm for pso}

\item{maxIter}{maxIter for pso}

\item{nCore}{number of core}
}
\value{
A list on design parameters and operating characteristics
}
\description{
PSOGO for maximizing power design with dual boundaries
}
