% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gbop2.R
\name{summary.gbop2}
\alias{summary.gbop2}
\title{Summary function
Summary function for gbop2 objects}
\usage{
\method{summary}{gbop2}(object, ...)
}
\arguments{
\item{object}{GBOP2_maxP_dualE GBOP2_maxP_singleE GBOP2_maxP_TE GBOP2_minSS_dualE GBOP2_minSS_singleE GBOP2_minSS_TE}

\item{...}{ignored arguments}
}
\value{
A summary table
}
\description{
Summary function
Summary function for gbop2 objects
}
\examples{
\donttest{
design <- GBOP2_minSS_singleE(
  design = "optimal", 
  unified.u = 1, 
  nlooks = 1, 
  p0 = 0.2, 
  p1 = 0.4, 
  err1 = 0.05, 
  minPower = 0.8, 
  weight = 1, 
  maxPatients = 25, 
  Nmin_cohort1 = 10, 
  Nmin_increase = 5, 
  pso_method = "default", 
  nParallel = 1, 
  seed = 1024, 
  nSwarm = 64, 
  maxIter = 200
)

summary(design)
}

}
