% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBOP2_minSS_singleE.gbop2.R
\name{GBOP2_minSS_singleE}
\alias{GBOP2_minSS_singleE}
\title{PSOGO: Optimal/Minimax design with single boundary}
\usage{
GBOP2_minSS_singleE(
  design = "optimal",
  unified.u = 1,
  nlooks = 1,
  b1n = 0.2,
  b1a = 0.4,
  err1 = 0.05,
  nParallel = 3,
  minPower = 0.8,
  weight = 1,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  pso_method = "all",
  seed = 1024,
  nSwarm = 64,
  maxIter = 200,
  nCore = 4
)
}
\arguments{
\item{design}{choose from "optimal", "minimax", or "unified"}

\item{unified.u}{specify when design = "unified", u in zero to one}

\item{nlooks}{number of interim looks}

\item{b1n}{Null hypothesis response rate}

\item{b1a}{Alternative hypothesis response rate}

\item{err1}{Type I error rate}

\item{nParallel}{number of pso ensemble}

\item{minPower}{power}

\item{weight}{weight of sample size under null}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{pso_method}{"all" for using three distinct pso, otherwise indicate single pso method}

\item{seed}{Random seed for reproducibility}

\item{nSwarm}{nSwarm for pso}

\item{maxIter}{maxIter for pso}

\item{nCore}{number of core}
}
\value{
A list on design parameters and operating characteristics
}
\description{
This function implements PSOGO to find an optimal or minimax design with single boundary.
}
\examples{
GBOP2_minSS_singleE(
  design = "optimal", 
  unified.u = 1, 
  nlooks = 1, 
  b1n = 0.2, 
  b1a = 0.4, 
  err1 = 0.05, 
  nParallel = 3, 
  minPower = 0.8, 
  weight = 1, 
  maxPatients = 25, 
  Nmin_cohort1 = 10, 
  Nmin_increase = 5, 
  pso_method = "all", 
  seed = 1024, 
  nSwarm = 64, 
  maxIter = 200, 
  nCore = 8
)


}
