% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBOP2_minSS_TE.gbop2.R
\name{GBOP2_minSS_TE}
\alias{GBOP2_minSS_TE}
\title{PSOGO: Optimal/Minimax design with efficacy and toxicity boundaries}
\usage{
GBOP2_minSS_TE(
  design = "optimal",
  unified.u = 1,
  nlooks = 1,
  skip_efficacy = NULL,
  skip_toxicity = NULL,
  maxPatients = 50,
  Nmin_cohort1 = 10,
  Nmin_increase = 5,
  nParallel = 3,
  e1n = 0.3,
  e2n = 0.4,
  e3n = 0.2,
  e1a = 0.6,
  e2a = 0.2,
  e3a = 0.15,
  err_eff = 0.1,
  err_tox = 0.1,
  err_all = 0.05,
  power_eff = 0.8,
  power_tox = 0.8,
  power_all = 0.8,
  pso_method = "all",
  seed = 5321,
  nSwarm = 32,
  maxIter = 100,
  nCore = 4
)
}
\arguments{
\item{design}{choose from "optimal", "minimax", or "unified"}

\item{unified.u}{specify when design = "unified", u in zero to one}

\item{nlooks}{number of interim looks}

\item{skip_efficacy}{default is NULL, indicate skip efficacy as 1 and not skip as 0 in a vector}

\item{skip_toxicity}{default is NULL, indicate skip toxicity as 1 and not skip as 0 in a vector}

\item{maxPatients}{maximum number of patients}

\item{Nmin_cohort1}{minimum number of first cohort}

\item{Nmin_increase}{minimum number of increase in each cohort}

\item{nParallel}{number of pso ensemble}

\item{e1n}{H0 for efficacy}

\item{e2n}{H0 for toxicity}

\item{e3n}{H0 for Eff and Tox}

\item{e1a}{H1 for efficacy}

\item{e2a}{H1 for toxicity}

\item{e3a}{H1 for Eff and Tox}

\item{err_eff}{Type I error rate: Efficacious but toxic}

\item{err_tox}{Type I error rate: Safe but futile}

\item{err_all}{Type I error rate: Futile and toxic}

\item{power_eff}{power: Efficacious but toxic}

\item{power_tox}{power: Safe but futile}

\item{power_all}{power: Futile and toxic}

\item{pso_method}{"all" for using three distinct pso, otherwise indicate single pso method}

\item{seed}{Random seed for reproducibility}

\item{nSwarm}{nSwarm in PSO}

\item{maxIter}{maxIter in PSO}

\item{nCore}{number of core}
}
\value{
A list on design parameters and operating characteristics
}
\description{
This function implements PSOGO to find an optimal or minimax design with efficacy and toxicity boundaries.
}
\examples{
GBOP2_minSS_TE(
  design = "optimal", 
  unified.u = 1, 
  nlooks = 1, 
  skip_efficacy = NULL, 
  skip_toxicity = NULL, 
  maxPatients = 25, 
  Nmin_cohort1 = 10, 
  Nmin_increase = 5, 
  nParallel = 3, 
  e1n = 0.3, 
  e2n = 0.4, 
  e3n = 0.2, 
  e1a = 0.6, 
  e2a = 0.2, 
  e3a = 0.15, 
  err_eff = 0.1, 
  err_tox = 0.1, 
  err_all = 0.05, 
  power_eff = 0.8, 
  power_tox = 0.8, 
  power_all = 0.8, 
  pso_method = "all", 
  seed = 5321, 
  nSwarm = 32, 
  maxIter = 100, 
  nCore = 8
)

 
}
