% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gbop2.R
\name{summary.gbop2}
\alias{summary.gbop2}
\title{summary function for PSOdesign, PSO_power,PSOGO_Design, PSOGO_power, PSODesign_dual,PSOPower_dual, PSOGDesign_Dual, PSOGPower_dual, PSODesign_TE, PSOGO_design_TE}
\usage{
\method{summary}{gbop2}(object, ...)
}
\arguments{
\item{object}{PSOdesign, PSO_power,PSOGO_Design, PSOGO_power, PSODesign_dual,PSOPower_dual, PSOGDesign_Dual, PSOGPower_dual, PSODesign_TE, PSOGO_design_TE}

\item{...}{ignored arguments}
}
\value{
A summary table
}
\description{
summary function for PSOdesign, PSO_power,PSOGO_Design, PSOGO_power, PSODesign_dual,PSOPower_dual, PSOGDesign_Dual, PSOGPower_dual, PSODesign_TE, PSOGO_design_TE
}
