% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gbop2.R
\name{summary.gbop2}
\alias{summary.gbop2}
\title{Summary function
Summary function for gbop2 objects}
\usage{
\method{summary}{gbop2}(object, ...)
}
\arguments{
\item{object}{GBOP2_maxP_dualE GBOP2_maxP_singleE GBOP2_maxP_TE GBOP2_minSS_dualE GBOP2_minSS_singleE GBOP2_minSS_TE PSO_design PSO_design_dual PSO_design_TE PSO_power PSO_power_dual PSO_power_TE}

\item{...}{ignored arguments}
}
\value{
A summary table
}
\description{
Summary function
Summary function for gbop2 objects
}
\examples{
summary(GBOP2_maxP_singleE(
  nlooks = 1, b1n = 0.2, b1a = 0.4, err1 = 0.05, 
  nParallel = 3, minPower = 0.8, totalPatients = 50, 
  Nmin_cohort1 = 10, Nmin_increase = 5
))

}
