% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gbop2.R
\name{summary.gbop2}
\alias{summary.gbop2}
\title{summary function for GBOP2_minSS_single, GBOP2_maxP_single, GBOP2_minSS_dual, PSO_design, PSO_power, PSO_design_dual,PSO_power_dual, PSO_design_TE, PSO_power_TE}
\usage{
\method{summary}{gbop2}(object, ...)
}
\arguments{
\item{object}{GBOP2_minSS_single, GBOP2_maxP_single, GBOP2_minSS_dual, PSO_design, PSO_power, PSO_design_dual,PSO_power_dual, PSO_design_TE, PSO_power_TE}

\item{...}{ignored arguments}
}
\value{
A summary table
}
\description{
summary function for GBOP2_minSS_single, GBOP2_maxP_single, GBOP2_minSS_dual, PSO_design, PSO_power, PSO_design_dual,PSO_power_dual, PSO_design_TE, PSO_power_TE
}
